# Front/personas/src/app/api/models/personas.ts
/* tslint:disable */
/* eslint-disable */
export interface Personas {
  apellidos?: null | string;
  email?: null | string;
  fechaCreacion?: null | string;
  identificador?: number;
  noIdentificacion?: null | string;
  nombres?: null | string;
  tipoIdentificacion?: null | string;
}